Don't delete completed tasks
Reminders will delete all of a list's items when the list is removed by
the user, regardless of their status. However it is preferable to keep a
record of completed tasks in TW for burndowns and such.

After this change, deleting a list in Reminders should only mark
incomplete tasks as deleted in TW.